+++ src/main/java/org/summerboot/jexpress/annotation/Service.java
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2005-2026 Du Law Office - jExpress, The Summer Boot Framework Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://apache.org
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package org.summerboot.jexpress.annotation;
+
+import com.google.inject.BindingAnnotation;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author Changski Tie Zheng Zhang 张铁铮, 魏泽北, 杜旺财, 杜富贵
+ */
+@Target(value = {ElementType.TYPE, ElementType.PARAMETER, ElementType.METHOD})
+@Retention(value = RetentionPolicy.RUNTIME)
+@Documented
+@BindingAnnotation
+public @interface Service {
+
+    String NOT_TAGGED = "";
+    String NOT_NAMED = "";
+
+    Class[] binding() default {};
+
+    String AlternativeName() default NOT_TAGGED;
+
+    String named() default NOT_NAMED;
+
+    ChannelHandlerType type() default ChannelHandlerType.unknown;
+
+    enum ChannelHandlerType {
+        ReadIdle, WriteIdle, FileUpload, Websocket, Ping, Business, unknown
+    }
+}