+++ src/main/java/org/summerboot/jexpress/boot/BootPOI.java
@@ -16,7 +16,7 @@
  */
 package org.summerboot.jexpress.boot;
 
-import org.summerboot.jexpress.boot.annotation.Unique;
+import org.summerboot.jexpress.annotation.validation.Unique;
 
 /**
  * @author Changski Tie Zheng Zhang 张铁铮, 魏泽北, 杜旺财, 杜富贵