+++ src/main/java/org/summerboot/jexpress/boot/config/BootConfig.java
@@ -22,16 +22,17 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.summerboot.jexpress.annotation.config.Config;
+import org.summerboot.jexpress.annotation.config.ConfigFilename;
+import org.summerboot.jexpress.annotation.config.ConfigHeader;
 import org.summerboot.jexpress.boot.BootConstant;
-import org.summerboot.jexpress.boot.config.annotation.Config;
-import org.summerboot.jexpress.boot.config.annotation.ConfigHeader;
-import org.summerboot.jexpress.boot.config.annotation.ImportResource;
-import org.summerboot.jexpress.webserver.netty.AbortPolicyWithReport;
 import org.summerboot.jexpress.security.EncryptorUtil;
 import org.summerboot.jexpress.util.ApplicationUtil;
 import org.summerboot.jexpress.util.BeanUtil;
 import org.summerboot.jexpress.util.ReflectionUtil;
 import org.summerboot.jexpress.util.concurrent.EmptyBlockingQueue;
+import org.summerboot.jexpress.util.concurrent.NamedDefaultThreadFactory;
+import org.summerboot.jexpress.webserver.netty.AbortPolicyWithReport;
 
 import javax.net.ssl.KeyManagerFactory;
 import javax.net.ssl.TrustManagerFactory;
@@ -447,7 +448,7 @@ public abstract class BootConfig implements JExpressConfig {
         }
 
         String namespace = "";
-        ImportResource ir = (ImportResource) configClass.getAnnotation(ImportResource.class);
+        ConfigFilename ir = (ConfigFilename) configClass.getAnnotation(ConfigFilename.class);
         if (ir != null) {
             namespace = ir.namespace();
         }