+++ src/main/java/org/summerboot/jexpress/integration/cache/AuthTokenCache.java
@@ -19,7 +19,7 @@ package org.summerboot.jexpress.integration.cache;
 import org.summerboot.jexpress.boot.BootErrorCode;
 import org.summerboot.jexpress.controller.Err;
 import org.summerboot.jexpress.controller.authenticate.Caller;
-import org.summerboot.jexpress.integration.healthcheck.HealthChecker;
+import org.summerboot.jexpress.integration.HealthChecker;
 
 import java.util.ArrayList;
 import java.util.List;